djrest_base_requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-seed==0.2.2
djangorestframework
Faker==5.0.1
idna==2.10
Pillow==8.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
